Exception Handling for Client Requests on Orders

+++ src/main/java/com/example/demo/service/OrderService.java
@@ -8,8 +8,12 @@ import com.example.demo.model.UserOrders;
 import com.example.demo.model.projections.OrderDetailsView;
 import com.example.demo.repository.CartItemsRepository;
 import com.example.demo.repository.OrderRepository;
+import com.example.demo.repository.PersonRepository;
 import com.example.demo.repository.UserOrdersRepository;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+
+import javax.persistence.EntityNotFoundException;
 
 @Service
 public class OrderService implements OrderItemsDao {
@@ -17,25 +21,35 @@ public class OrderService implements OrderItemsDao {
     private OrderRepository orderRepository;
     private CartItemsRepository cartItemsRepository;
     private UserOrdersRepository userOrdersRepository;
+    private PersonRepository personRepository;
     
-    public OrderService(OrderRepository orderRepository, CartItemsRepository cartItemsRepository, UserOrdersRepository userOrdersRepository) {
+    public OrderService(OrderRepository orderRepository, CartItemsRepository cartItemsRepository, UserOrdersRepository userOrdersRepository, PersonRepository personRepository) {
         this.orderRepository = orderRepository;
         this.cartItemsRepository = cartItemsRepository;
         this.userOrdersRepository = userOrdersRepository;
+        this.personRepository = personRepository;
     }
 
     @Override
     public void save(long personId) {
-        Person person = new Person();
-        person.setId(personId);
-        Iterable<CartItems> cartItems = this.cartItemsRepository.findCartItemsByPerson(person);
-        UserOrders userOrder = new UserOrders(person);
-        UserOrders savedUserOrder = this.userOrdersRepository.save(userOrder);
-        for (CartItems cartItem : cartItems) {
-            OrderItems orderItem = new OrderItems(cartItem.getProduct(), cartItem.getQuantity(), savedUserOrder);
-            OrderItems savedOrderItem = this.orderRepository.save(orderItem);
+        if (personRepository.findById(personId).isPresent()) {
+            Person person = new Person();
+            person.setId(personId);
+            Iterable<CartItems> cartItems = cartItemsRepository.findCartItemsByPerson(person);
+            if (cartItems.iterator().hasNext()) {
+                UserOrders userOrder = new UserOrders(person);
+                UserOrders savedUserOrder = userOrdersRepository.save(userOrder);
+                for (CartItems cartItem : cartItems) {
+                    OrderItems orderItem = new OrderItems(cartItem.getProduct(), cartItem.getQuantity(), savedUserOrder);
+                    OrderItems savedOrderItem = orderRepository.save(orderItem);
+                }
+                this.cartItemsRepository.deleteByPerson(person);
+            } else {
+                throw new EntityNotFoundException();
+            }
+        } else {
+            throw new DataIntegrityViolationException("User doesnt exist with that ID");
         }
-        this.cartItemsRepository.deleteByPerson(person);
     }
 
     @Override
